wire file tree with editor
