+++ src/app/code-editor-demo/code-editor-demo.component.html
@@ -10,52 +10,40 @@
         {{ theme.name }}
       </button>
     </mat-menu>
-
-    <div>
-      <mat-form-field>
-        <mat-select
-          [(value)]="selectedModel"
-          (selectionChange)="onDemoChanged($event)">
-          <mat-option *ngFor="let model of demos" [value]="model">
-            {{ model.language }}
-          </mat-option>
-        </mat-select>
-      </mat-form-field>
-    </div>
   </mat-toolbar>
 
   <div fxLayout fxFlex>
     <div fxFlex="20">
-      <div>
+      <div class="file-tree-container">
         <mat-tree [dataSource]="nestedDataSource" [treeControl]="nestedTreeControl" class="file-tree">
-            <mat-tree-node *matTreeNodeDef="let node" matTreeNodeToggle>
-              <li class="mat-tree-node">
-                <button mat-icon-button disabled></button>
-                {{node.filename}}:  {{node.type}}
-              </li>
-            </mat-tree-node>
+          <mat-tree-node *matTreeNodeDef="let node" matTreeNodeToggle>
+            <li class="mat-tree-node">
+              <button mat-icon-button disabled></button>
+              <span class="file-node-label" (click)="selectNode(node)">{{node.filename}}.{{node.type}}</span>
+            </li>
+          </mat-tree-node>
 
-            <mat-nested-tree-node *matTreeNodeDef="let node; when: hasNestedChild">
-              <li>
-                <div class="mat-tree-node">
-                  <button mat-icon-button matTreeNodeToggle
-                          [attr.aria-label]="'toggle ' + node.filename">
-                    <mat-icon class="mat-icon-rtl-mirror">
-                      {{nestedTreeControl.isExpanded(node) ? 'expand_more' : 'chevron_right'}}
-                    </mat-icon>
-                  </button>
-                  <span class="file-node-label" (click)="nestedTreeControl.toggle(node)">{{node.filename}}</span>
-                </div>
-                <ul [class.file-tree-invisible]="!nestedTreeControl.isExpanded(node)">
-                  <ng-container matTreeNodeOutlet></ng-container>
-                </ul>
-              </li>
-            </mat-nested-tree-node>
-          </mat-tree>
-        </div>
+          <mat-nested-tree-node *matTreeNodeDef="let node; when: hasNestedChild">
+            <li>
+              <div class="mat-tree-node">
+                <button mat-icon-button matTreeNodeToggle
+                        [attr.aria-label]="'toggle ' + node.filename">
+                  <mat-icon class="mat-icon-rtl-mirror">
+                    {{nestedTreeControl.isExpanded(node) ? 'expand_more' : 'chevron_right'}}
+                  </mat-icon>
+                </button>
+                <span class="file-node-label" (click)="nestedTreeControl.toggle(node)">{{node.filename}}</span>
+              </div>
+              <ul [class.file-tree-invisible]="!nestedTreeControl.isExpanded(node)">
+                <ng-container matTreeNodeOutlet></ng-container>
+              </ul>
+            </li>
+          </mat-nested-tree-node>
+        </mat-tree>
+      </div>
     </div>
     <div fxFlex="80">
-      <ngs-code-editor
+      <ngs-code-editor *ngIf="selectedModel"
         [codeModel]="selectedModel"
         [theme]="activeTheme"
         [options]="options"
