+++ src/app/code-editor-demo/code-editor-demo.component.scss
@@ -13,7 +13,21 @@
   }
 }
 
+.file-tree-container {
+  height: calc(100vh - 64px);
+  overflow: hidden;
+}
+
 .file-tree {
+  display: flex;
+  flex-direction: column;
+  flex: 1;
+  height: 100%;
+  max-height: 100%;
+  overflow: hidden;
+  min-height: 0;
+  overflow: auto;
+
   &-invisible {
     display: none;
   }
