+++ src/app/code-editor-demo/code-editor-demo.component.ts
@@ -5,7 +5,6 @@ import {
   ViewEncapsulation,
   OnInit
 } from '@angular/core';
-import { MatSelectChange } from '@angular/material/select';
 import { CodeModel } from '@ngstack/code-editor';
 import { NestedTreeControl } from '@angular/cdk/tree';
 import { MatTreeNestedDataSource } from '@angular/material/tree';
@@ -29,56 +28,8 @@ export class CodeEditorDemoComponent implements OnInit {
     { name: 'High Contrast Dark', value: 'hc-black' }
   ];
 
-  demos: Array<CodeModel> = [
-    {
-      language: 'typescript',
-      uri: 'main.ts',
-      dependencies: ['@ngstack/translate', '@ngstack/code-editor'],
-      value: `
-        // TypeScript Example
-        import { TranslateModule, TranslateService } from '@ngstack/translate';
-        import { CodeEditorModule } from '@ngstack/code-editor';
-        import * as fs from 'fs';
-
-        export class MyClass {
-          constructor(translate: TranslateService) {
-
-          }
-        }
-      `
-    },
-    {
-      language: 'javascript',
-      uri: 'main.js',
-      dependencies: ['@types/node'],
-      value: `
-        // JavaScript Example
-        import * as fs from 'fs';
-
-        class Person {
-          greet() {
-            console.log('hello there');
-            fs.mkdir('folder');
-          }
-        }
-      `
-    },
-    {
-      language: 'json',
-      uri: 'main.json',
-      value: [
-        '{',
-        '    "$schema": "http://myserver/foo-schema.json",',
-        '    "p1": "v3",',
-        '    "p2": false',
-        '}'
-      ].join('\n')
-    }
-  ];
-
-  selectedModel = this.demos[2];
+  selectedModel: CodeModel = null;
   activeTheme = 'vs';
-  code = this.demos[2].value;
   readOnly = false;
 
   @ViewChild('file')
@@ -106,13 +57,9 @@ export class CodeEditorDemoComponent implements OnInit {
     // console.log('CODE', value);
   }
 
-  onLoadClicked() {
-    this.fileInput.nativeElement.click();
-  }
-
-  onDemoChanged(event: MatSelectChange) {
-    const model: CodeModel = event.value;
-    this.code = model.value;
+  selectNode(node: FileNode) {
+    console.log(node);
+    this.selectedModel = node.code;
   }
 
   ngOnInit() {}
