+++ src/app/code-editor-demo/file-node.ts
@@ -1,5 +1,9 @@
+import { CodeModel } from '@ngstack/code-editor/public_api';
+
 export class FileNode {
   children?: FileNode[];
   filename: string;
   type?: any;
+
+  code?: CodeModel;
 }
